Added option for change no of recommenations
