+++ UseCases/Recommendations_product_search/utils/sql_helper_func.py
@@ -1,0 +1,35 @@
+# teradata lib
+from teradataml.dataframe.copy_to import copy_to_sql
+
+#  SQL code
+# delete all the records from table
+def delete_emb_from_sql(table_name, eng):
+    qry = f"DELETE FROM {table_name}"
+    try:
+        eng.execute(qry)
+        print(f'All the records deleted from: {table_name}')
+    except:
+        pass
+
+    
+
+# only copy embeddings, if table exists then append
+def copy_emb_to_sql(table_name, tdf):
+    try:
+        copy_to_sql(tdf, table_name=table_name,primary_index='product_id', if_exists='append')
+    except:
+        pass
+
+def delete_and_copy_embeddings(table_name, tdf, eng):
+    # only delete records if table exists
+    qry = f"DELETE FROM {table_name}"
+    try:
+        eng.execute(qry)
+    except:
+        pass
+
+    # only insert records if table exists
+    try:
+        copy_to_sql(tdf, table_name=table_name,primary_index='product_id', if_exists='append')
+    except:
+        pass
